@@ -19,12 +19,16 @@
       "cweagans/composer-patches": true,
       "dealerdirect/phpcodesniffer-composer-installer": true,
       "johnpbloch/wordpress-core-installer": true,
-      "phpstan/extension-installer": true,
-      "brianhenryie/composer-phpstorm": true
+      "phpstan/extension-installer": true
     },
     "platform": {
-      "php": "8.0.30"
+      "php": "8.4"
     }
+  },
+  "autoload": {
+    "classmap":  [
+      "includes"
+    ]
   },
   "autoload-dev": {
       "psr-4": {
@@ -130,7 +134,7 @@
     }
   },
   "require": {
-    "php": ">=8.0",
+    "php": ">=8.4",
     "ext-bcmath": "*",
     "ext-intl": "*",
     "ext-json": "*",
