@@ -104,7 +104,7 @@ class Bitcoin_Address_Repository {
 
 	/**
 	 * @param string $post_staus
-	 * @param int $number_posts Defaults to WP_Query's max of 200.
+	 * @param int    $number_posts Defaults to WP_Query's max of 200.
 	 *
 	 * @return \WP_Post[]
 	 */
@@ -124,12 +124,22 @@ class Bitcoin_Address_Repository {
 	/**
 	 * @return \WP_Post[]
 	 */
-	public function get_assigned_bitcoin_addresses(): array {
+	public function get_assigned_bitcoin_addresses_wp_posts(): array {
 		return $this->get_bitcoin_address_posts( 'assigned' );
 	}
 
 	/**
-	 * Check do we have 1 assigned address.
+	 * @return Bitcoin_Address[]
+	 */
+	public function get_assigned_bitcoin_addresses(): array {
+		return array_map(
+			fn( \WP_Post $bitcoin_address_wp_post ) => new Bitcoin_Address( $bitcoin_address_wp_post->ID ),
+			$this->get_bitcoin_address_posts( 'assigned' )
+		);
+	}
+
+	/**
+	 * Check do we have at least 1 assigned address, i.e. an address waiting for transactions.
 	 */
 	public function has_assigned_bitcoin_addresses(): bool {
 		$assigned_addresses = $this->get_bitcoin_address_posts( 'assigned', 1 );
