@@ -13,7 +13,7 @@ use BrianHenryIE\WP_Bitcoin_Gateway\API\API;
 /**
  * Verifies the plugin has been instantiated and added to PHP's $GLOBALS variable.
  */
-class Plugin_Integration_Test extends \Codeception\TestCase\WPTestCase {
+class Plugin_Integration_Test extends \lucatume\WPBrowser\TestCase\WPTestCase {
 
 	/**
 	 * Test the main plugin object is added to PHP's GLOBALS and that it is the correct class.
