@@ -3,11 +3,13 @@
 namespace BrianHenryIE\WP_Bitcoin_Gateway\Integrations\WooCommerce;
 
 use BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler\Background_Jobs;
+use BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler\Background_Jobs_Actions_Interface;
+use lucatume\WPBrowser\TestCase\WPTestCase;
 
 /**
  * @coversNothing
  */
-class Order_Integration_Test extends \Codeception\TestCase\WPTestCase {
+class Order_Integration_Test extends WPTestCase {
 
 	/**
 	 * It seemed the check-unpaid-order job was not being cancelled when the order was paid, but the test
@@ -22,7 +24,7 @@ class Order_Integration_Test extends \Codeception\TestCase\WPTestCase {
 		$order->set_status( 'on-hold' );
 		$order->save();
 
-		$hook = Background_Jobs::CHECK_UNPAID_ORDER_HOOK;
+		$hook = Background_Jobs_Actions_Interface::CHECK_ASSIGNED_ADDRESSES_TRANSACTIONS_HOOK;
 		$args = array( 'order_id' => $order_id );
 
 		assert( as_has_scheduled_action( $hook, $args ) );
