@@ -3,6 +3,7 @@
 namespace BrianHenryIE\WP_Bitcoin_Gateway\API;
 
 use BrianHenryIE\ColorLogger\ColorLogger;
+use BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler\Background_Jobs;
 use BrianHenryIE\WP_Bitcoin_Gateway\API\Addresses\Bitcoin_Address;
 use BrianHenryIE\WP_Bitcoin_Gateway\API\Addresses\Bitcoin_Address_Repository;
 use BrianHenryIE\WP_Bitcoin_Gateway\API\Addresses\Bitcoin_Wallet_Factory;
@@ -35,8 +36,8 @@ class API_Unit_Test extends \Codeception\Test\Unit {
 		$logger   = new ColorLogger();
 		$settings = $this->makeEmpty( Settings_Interface::class );
 
-		$bitcoin_wallet_factory  = $this->makeEmpty( Bitcoin_Wallet_Factory::class );
-		$bitcoin_address_factory = $this->makeEmpty( Bitcoin_Address_Repository::class );
+		$bitcoin_wallet_factory     = $this->makeEmpty( Bitcoin_Wallet_Factory::class );
+		$bitcoin_address_repository = $this->makeEmpty( Bitcoin_Address_Repository::class );
 
 		$transaction = self::makeEmpty(
 			Transaction_Interface::class,
@@ -60,7 +61,9 @@ class API_Unit_Test extends \Codeception\Test\Unit {
 		$generate_address_api = $this->makeEmpty( Generate_Address_API_Interface::class );
 		$exchange_rate_api    = $this->makeEmpty( Exchange_Rate_API_Interface::class );
 
-		$sut = new API( $settings, $logger, $bitcoin_wallet_factory, $bitcoin_address_factory, $blockchain_api, $generate_address_api, $exchange_rate_api );
+		$background_jobs = $this->makeEmpty( Background_Jobs::class );
+
+		$sut = new API( $settings, $logger, $bitcoin_wallet_factory, $bitcoin_address_repository, $blockchain_api, $generate_address_api, $exchange_rate_api, $background_jobs );
 
 		$address = self::make(
 			Bitcoin_Address::class,
@@ -81,16 +84,16 @@ class API_Unit_Test extends \Codeception\Test\Unit {
 	 */
 	public function test_get_exchange_rate_already_stored(): void {
 
-		$logger                  = new ColorLogger();
-		$settings                = $this->makeEmpty( Settings_Interface::class );
-		$bitcoin_wallet_factory  = $this->makeEmpty( Bitcoin_Wallet_Factory::class );
-		$bitcoin_address_factory = $this->makeEmpty( Bitcoin_Address_Repository::class );
-		$blockchain_api          = $this->makeEmpty( Blockchain_API_Interface::class );
-		$generate_address_api    = $this->makeEmpty( Generate_Address_API_Interface::class );
-		$exchange_rate_api       = $this->makeEmpty( Exchange_Rate_API_Interface::class );
+		$logger                     = new ColorLogger();
+		$settings                   = $this->makeEmpty( Settings_Interface::class );
+		$bitcoin_wallet_factory     = $this->makeEmpty( Bitcoin_Wallet_Factory::class );
+		$bitcoin_address_repository = $this->makeEmpty( Bitcoin_Address_Repository::class );
+		$blockchain_api             = $this->makeEmpty( Blockchain_API_Interface::class );
+		$generate_address_api       = $this->makeEmpty( Generate_Address_API_Interface::class );
+		$exchange_rate_api          = $this->makeEmpty( Exchange_Rate_API_Interface::class );
+		$background_jobs            = $this->makeEmpty( Background_Jobs::class );
 
-		$sut = new API( $settings, $logger, $bitcoin_wallet_factory, $bitcoin_address_factory, $blockchain_api, $generate_address_api, $exchange_rate_api );
-
+		$sut = new API( $settings, $logger, $bitcoin_wallet_factory, $bitcoin_address_repository, $blockchain_api, $generate_address_api, $exchange_rate_api, $background_jobs );
 		\WP_Mock::userFunction(
 			'get_transient',
 			array(
@@ -117,13 +120,13 @@ class API_Unit_Test extends \Codeception\Test\Unit {
 	 */
 	public function test_get_exchange_rate_not_already_stored(): void {
 
-		$logger                  = new ColorLogger();
-		$settings                = $this->makeEmpty( Settings_Interface::class );
-		$bitcoin_wallet_factory  = $this->makeEmpty( Bitcoin_Wallet_Factory::class );
-		$bitcoin_address_factory = $this->makeEmpty( Bitcoin_Address_Repository::class );
-		$blockchain_api          = $this->makeEmpty( Blockchain_API_Interface::class );
-		$generate_address        = $this->makeEmpty( Generate_Address_API_Interface::class );
-		$exchange_rate_api       = $this->makeEmpty(
+		$logger                     = new ColorLogger();
+		$settings                   = $this->makeEmpty( Settings_Interface::class );
+		$bitcoin_wallet_factory     = $this->makeEmpty( Bitcoin_Wallet_Factory::class );
+		$bitcoin_address_repository = $this->makeEmpty( Bitcoin_Address_Repository::class );
+		$blockchain_api             = $this->makeEmpty( Blockchain_API_Interface::class );
+		$generate_address           = $this->makeEmpty( Generate_Address_API_Interface::class );
+		$exchange_rate_api          = $this->makeEmpty(
 			Exchange_Rate_API_Interface::class,
 			array(
 				'get_exchange_rate' => Expected::once(
@@ -135,7 +138,9 @@ class API_Unit_Test extends \Codeception\Test\Unit {
 			)
 		);
 
-		$sut = new API( $settings, $logger, $bitcoin_wallet_factory, $bitcoin_address_factory, $blockchain_api, $generage_address, $exchange_rate_api );
+		$background_jobs = $this->makeEmpty( Background_Jobs::class );
+
+		$sut = new API( $settings, $logger, $bitcoin_wallet_factory, $bitcoin_address_repository, $blockchain_api, $generate_address, $exchange_rate_api, $background_jobs );
 
 		\WP_Mock::userFunction(
 			'get_transient',
