@@ -7,11 +7,13 @@
 namespace BrianHenryIE\WP_Bitcoin_Gateway;
 
 use BrianHenryIE\ColorLogger\ColorLogger;
+use BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler\API_Background_Jobs_Interface;
 use BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler\Background_Jobs;
 use BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler\Background_Jobs_Actions_Interface;
 use BrianHenryIE\WP_Bitcoin_Gateway\Action_Scheduler\Background_Jobs_Scheduling_Interface;
 use BrianHenryIE\WP_Bitcoin_Gateway\Admin\Plugins_Page;
 use BrianHenryIE\WP_Bitcoin_Gateway\Admin\Register_List_Tables;
+use BrianHenryIE\WP_Bitcoin_Gateway\API\API;
 use BrianHenryIE\WP_Bitcoin_Gateway\Frontend\Frontend_Assets;
 use BrianHenryIE\WP_Bitcoin_Gateway\Integrations\Woo_Cancel_Abandoned_Order\Woo_Cancel_Abandoned_Order;
 use BrianHenryIE\WP_Bitcoin_Gateway\lucatume\DI52\Container;
@@ -59,6 +61,13 @@ class BH_WP_Bitcoin_Gateway_Unit_Test extends \Codeception\Test\Unit {
 		);
 		$container->bind( Settings_Interface::class, $settings );
 		$container->bind( LoggerInterface::class, ColorLogger::class );
+
+		$container->bind(
+			API_Background_Jobs_Interface::class,
+			function () {
+				return self::makeEmpty( API_Background_Jobs_Interface::class );
+			}
+		);
 
 		$container->bind( Background_Jobs_Scheduling_Interface::class, Background_Jobs::class );
 		$container->bind( Background_Jobs_Actions_Interface::class, Background_Jobs::class );
@@ -212,13 +221,23 @@ class BH_WP_Bitcoin_Gateway_Unit_Test extends \Codeception\Test\Unit {
 	public function test_define_action_scheduler_hooks(): void {
 
 		\WP_Mock::expectActionAdded(
-			Background_Jobs::CHECK_ASSIGNED_ADDRESSES_TRANSACTIONS_HOOK,
-			array( new AnyInstance( Background_Jobs::class ), 'check_unpaid_order' )
+			Background_Jobs_Actions_Interface::GENERATE_NEW_ADDRESSES_HOOK,
+			array( new AnyInstance( Background_Jobs::class ), 'generate_new_addresses' )
 		);
 
 		\WP_Mock::expectActionAdded(
-			Background_Jobs::GENERATE_NEW_ADDRESSES_HOOK,
-			array( new AnyInstance( Background_Jobs::class ), 'generate_new_addresses' )
+			Background_Jobs_Actions_Interface::CHECK_NEW_ADDRESSES_TRANSACTIONS_HOOK,
+			array( new AnyInstance( Background_Jobs::class ), 'check_new_addresses_for_transactions' )
+		);
+
+		\WP_Mock::expectActionAdded(
+			Background_Jobs_Actions_Interface::CHECK_ASSIGNED_ADDRESSES_TRANSACTIONS_HOOK,
+			array( new AnyInstance( Background_Jobs::class ), 'check_assigned_addresses_for_transactions' )
+		);
+
+		\WP_Mock::expectActionAdded(
+			Background_Jobs_Actions_Interface::CHECK_FOR_ASSIGNED_ADDRESSES_HOOK,
+			array( new AnyInstance( Background_Jobs::class ), 'ensure_schedule_repeating_actions' )
 		);
 
 		$app = new BH_WP_Bitcoin_Gateway( $this->get_container() );
