@@ -18,8 +18,8 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 	 * @covers ::__construct
 	 */
 	public function test_get_id(): void {
-		$bitcoin_address_mock    = self::make( Bitcoin_Address::class );
-		$bitcoin_address_factory = self::make(
+		$bitcoin_address_mock       = self::make( Bitcoin_Address::class );
+		$bitcoin_address_repository = self::make(
 			Bitcoin_Address_Repository::class,
 			array(
 				'get_post_id_for_address' => Expected::once( 123 ),
@@ -35,7 +35,7 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 		$order->add_meta_data( Order::LAST_CHECKED_META_KEY, new \DateTime(), true );
 		$order_id = $order->save();
 
-		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_factory );
+		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_repository );
 
 		$result = $sut->get_id();
 
@@ -46,11 +46,11 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 	 * @covers ::get_address
 	 */
 	public function test_get_address(): void {
-		$bitcoin_address_mock    = self::make(
+		$bitcoin_address_mock       = self::make(
 			Bitcoin_Address::class,
 			array( 'get_raw_address' => Expected::once( 'success' ) )
 		);
-		$bitcoin_address_factory = self::make(
+		$bitcoin_address_repository = self::make(
 			Bitcoin_Address_Repository::class,
 			array(
 				'get_post_id_for_address' => Expected::once( 123 ),
@@ -66,7 +66,7 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 		$order->add_meta_data( Order::LAST_CHECKED_META_KEY, new \DateTime(), true );
 		$order_id = $order->save();
 
-		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_factory );
+		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_repository );
 
 		$result = $sut->get_address();
 
@@ -77,8 +77,8 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 	 * No covers because it uses a __call @method.
 	 */
 	public function test_is_paid(): void {
-		$bitcoin_address_mock    = self::make( Bitcoin_Address::class );
-		$bitcoin_address_factory = self::make(
+		$bitcoin_address_mock       = self::make( Bitcoin_Address::class );
+		$bitcoin_address_repository = self::make(
 			Bitcoin_Address_Repository::class,
 			array(
 				'get_post_id_for_address' => Expected::once( 123 ),
@@ -96,7 +96,7 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 
 		add_filter( 'woocommerce_order_is_paid', '__return_true' );
 
-		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_factory );
+		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_repository );
 
 		$result = $sut->is_paid();
 
@@ -107,8 +107,8 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 	 * @covers ::set_last_checked_time
 	 */
 	public function test_set_last_checked_time(): void {
-		$bitcoin_address_mock    = self::make( Bitcoin_Address::class );
-		$bitcoin_address_factory = self::make(
+		$bitcoin_address_mock       = self::make( Bitcoin_Address::class );
+		$bitcoin_address_repository = self::make(
 			Bitcoin_Address_Repository::class,
 			array(
 				'get_post_id_for_address' => Expected::once( 123 ),
@@ -124,7 +124,7 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 		$order->add_meta_data( Order::LAST_CHECKED_META_KEY, new DateTimeImmutable(), true );
 		$order_id = $order->save();
 
-		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_factory );
+		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_repository );
 
 		// 946684800 is Y2K.
 		$last_checked = DateTimeImmutable::createFromFormat( 'U', '946684800' );
@@ -144,8 +144,8 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 	 * No covers because it uses a __call @method.
 	 */
 	public function test_get_status(): void {
-		$bitcoin_address_mock    = self::make( Bitcoin_Address::class );
-		$bitcoin_address_factory = self::make(
+		$bitcoin_address_mock       = self::make( Bitcoin_Address::class );
+		$bitcoin_address_repository = self::make(
 			Bitcoin_Address_Repository::class,
 			array(
 				'get_post_id_for_address' => Expected::once( 123 ),
@@ -162,7 +162,7 @@ class Bitcoin_Order_WPUnit_Test extends \lucatume\WPBrowser\TestCase\WPTestCase 
 		$order->add_meta_data( Order::LAST_CHECKED_META_KEY, new \DateTime(), true );
 		$order_id = $order->save();
 
-		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_factory );
+		$sut = new WC_Bitcoin_Order( $order, $bitcoin_address_repository );
 
 		$result = $sut->get_status();
 
